Add verbose log message pointing to sibling git dir
Adds a log message showing the location of the directory where the actual git prepare is running (but only if `--verbose` is passed).

Might be useful for other users encountering git errors and not being able to reproduce/fix because they are not aware of the "sibling directory" being used. Like was the case here https://github.com/ef4/ember-cli-deploy-git/issues/20.